+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Enhancements
 - **Dashboard Changes**
   - Renamed "Research Highlights" to "Highlights"
+- **Learn Page Changes**
+  - Added Complex Sleep Apnea to the learn page
 - **Password Reset Changes**
   - Password reset emails are no longer sent to deleted users
   - Provided better message if the email address was not found